### 2020-02-04 #2069 @whedon
[PRE REVIEW]: BayesMFSurv: An R Package to Estimate Bayesian Split-Population Survival Models With (and Without) Misclassified Failure Events
**Submitting author:** @Nicolas-Schmidt (<a href="http://orcid.org/0000-0001-5083-5792">Nicolás Schmidt</a>)
**Repository:** <a href="https://github.com/Nicolas-Schmidt/BayesMFSurv" target ="_blank">https://github.com/Nicolas-Schmidt/BayesMFSurv</a>
**Version:** 0.2.0
