### 2020-02-06 a4ec68e
Update mcmcSurv.R
