### 2020-02-06 49d13c5
Update mcmcSurv.R
