### 2020-02-06 f37d8d1
Update mcmcSurv.R
