### 2020-02-11 #2069 @marcosvital
Ah, didn't know that, thanks @danielskatz ! :)

Hi, @Nicolas-Schmidt, I'll be the editor of your submission.

Please take a look at the article proof generated above, and check if everything seems to be ok.

On the meantime, we can find reviewers for you submission. We usually ask the authors for suggestions, so let me know if you have anyone in mind that could do this. You can check [this list of people](https://bit.ly/joss-reviewers) that have agreed to review for JOSS, and see if any of them may be suitable to review this submission - but feel free to suggest other people as well, ok?

And let me know if you need any help finding reviewers if you don't have anyone in mind.
