+++ paper.bib
@@ -111,7 +111,8 @@
   author    = {Maller, Ross A and Zhou, Xian},
   year      = {1996},
   publisher = {John Wiley \& Sons},
-  isbn      = {978-0-471-96201-4}
+  isbn      = {978-0-471-96201-4},
+  doi       = {10.1080/00401706.1998.10485509}
 }
 
 @manual{r-core,
