### 2020-02-24 #2069 @Sergio-Bejar
@danielskatz @kthyng @marcosvital , would it be possible to get an update on the status of our paper?
