### 2020-02-26 #2069 @Sergio-Bejar
@danielskatz @kthyng, apologies for bugging you about this submission.   But there's been no feedback/information from the assigned editor in the past 14 days -specifically since we proposed a set of potential reviewers.-   Other submissions assigned to the same editor seem to have the same issue (no response).  Is it possible to assign the piece to a different editor?  If not, is there anything we can do to help speed up (or at least move to the review stage) process?

Again, apologies for the multiple messages and thanks for your time.
