### 2020-02-27 #2069 @marcosvital
Hi, @Sergio-Bejar ! I'm terribly sorry about the delayed reply and for not being able to check in to follow your submission and keep it moving - I was unable to work on this over the last couple of weeks, and should have let the rest of the team know and assigned someone else. I'm truly sorry about this, and will note let that happen again.

I'll follow up calling the potential reviewers, ok, and keep constantly checking the status of your submission. But please let us know if you don't feel comfortable with me as an editor from this point on.
