### 2020-02-27 #2069 @marcosvital
Dear @andybega, @alletsee and @rdrr1990:

Would you be willing to review this submission for JOSS?

This is still a pre-review, once there are enough reviewers, we will move on to a new issue where the review will take place.

If you are not able, please let us know if you have recommendations for other reviewers.
