### 2020-03-16 #2164 @marcosvital
Dear @alletsee and @andybega, thank you very much for accepting review this submission for JOSS.
Even if you are not starting the review right now, please accept the invite, as it has an expiration date (there is a link under **Reviewer instructions & questions** and you should also get an email notification). Furthermore, please check the instructions and checklists above, and let me know if you need any assistance.

You can also tag @Nicolas-Schmidt and @Sergio-Bejar  if you need to ask specific questions about the submission.
