### 2020-03-19 #2164 @alletsee
Dear @Nicolas-Schmidt and @Sergio-Bejar,

Thanks for submitting the software and this paper - and congratulations for publishing the academic paper in PA!

I'm currently reviewing the package and have a few minor questions:
- Tests: I couldn't see any automated testing integrated in the paper, but this is part of the JOSS checklist. Did I overlook something? Or can we maybe use something as a surrogate for this? (@marcosvital : would replicating the coefficients from the main example be acceptable for this point?)
- Community Guidelines: I understand that this is meant as an addition to an academic paper, so I wouldn't expect a massive developer community starting to extend the software. However, I think it might be good if you would add a note how users can report issues/problems or seek support (e.g. by recommending to open an issue on Github). Is this something that you might consider adding?
- Documentation, #1: In "Buhaugetal_2009_JCR.Rd", you refer to a "Gates et al. (2006)" paper for which I cannot find a proper citation. Could you please add this? (Note that you can add a "Reference" section to Rd files, which might be a convenient and user-friendly way.)
- Documentation, #2: Super-tiny detail, but I just spotted this and it's super-quick to fix. In "RBS.Rd", there is a missing white space for "Bernhard,Michael," in the source.

I hope that the questions/suggestions above are helpful for you and not much of a burden.
(As a quick status update: I still have to read the software paper and test the code to confirm functionality.)

Best regards,
Max
