+++ paper.bib
@@ -139,6 +139,7 @@
   year      = {2018},
   volume    = {accept minor},
   url       = {http://arxiv.org/abs/1705.04584},
+  doi       = {10.18637/jss.v092.i09}
 }
 
 @Manual{wangetal-dynsurv,
