+++ DESCRIPTION
@@ -32,6 +32,7 @@ Encoding: UTF-8
 LazyData: true
 Depends:
     R (>= 3.5.0)
+BugReports: https://github.com/Nicolas-Schmidt/BayesMFSurv/issues
 URL: https://github.com/Nicolas-Schmidt/BayesMFSurv
 Imports:
 	MCMCpack,
