### 2020-03-19 #2164 @Nicolas-Schmidt
Dear @alletsee:

Thanks for your comments and questions.  They all are very helpful.  Below please find our response to your comments/queries.

1)   This is a good point.  We did not integrate any automatic testing into the paper due to space limitations.   But the README file of our package provides readers with examples that allow users to verify the functionality of the software.   We hope this is acceptable, and will be more than happy to add a sentence in the paper explaining where users can find such examples.
2)   We addressed this point by opening an issue on Github where users can report issues/problems or seek support. This can be found in the “DESCRIPTION” file on Github.
3)   Thanks for raising this issue.  We indeed omitted the DOI of the Gates et al (2006) paper in our original submission.
4)   We fixed the issue related to the missing white space for “Bernhard, Michael,” in the source.
5)    Further, the automatic compilation detected a missing DOI.  This has also been corrected as we added the aforementioned DOI.

Thanks again for your time and effort reviewing our submission!
