### 2020-03-20 Nicolas-Schmidt/BayesMFSurv#1 @andybega
R check error
Hi guys,

I tried rebuilding the package using the latest from GH (I cloned your repo), and:

1. Some of the documentation is out of date (I ran `devtools::document()` and there were some minor changes).
2. R check is throwing an error:

```
── Building ───────────────────────────────────────────────────────────── BayesMFSurv ──
Setting env vars:
● CFLAGS    : -Wall -pedantic -fdiagnostics-color=always
● CXXFLAGS  : -Wall -pedantic -fdiagnostics-color=always
● CXX11FLAGS: -Wall -pedantic -fdiagnostics-color=always
────────────────────────────────────────────────────────────────────────────────────────
✓  checking for file ‘/Users/andybega/Work/BayesMFSurv/DESCRIPTION’ ...
─  preparing ‘BayesMFSurv’:
✓  checking DESCRIPTION meta-information ...
─  cleaning src
─  looking to see if a ‘data/datalist’ file should be added
─  building ‘BayesMFSurv_0.2.0.tar.gz’

── Checking ───────────────────────────────────────────────────────────── BayesMFSurv ──
Setting env vars:
● _R_CHECK_CRAN_INCOMING_REMOTE_: FALSE
● _R_CHECK_CRAN_INCOMING_       : FALSE
● _R_CHECK_FORCE_SUGGESTS_      : FALSE
── R CMD check ─────────────────────────────────────────────────────────────────
─  using log directory ‘/Users/andybega/Work/BayesMFSurv.Rcheck’
─  using R version 3.6.1 (2019-07-05)
─  using platform: x86_64-apple-darwin15.6.0 (64-bit)
─  using session charset: UTF-8
─  using options ‘--no-manual --as-cran’
✓  checking for file ‘BayesMFSurv/DESCRIPTION’ ...
─  checking extension type ... Package
─  this is package ‘BayesMFSurv’ version ‘0.2.0’
─  package encoding: UTF-8
✓  checking package namespace information ...
E  checking package dependencies (3.5s)
   Namespace dependency not required: ‘RcppArmadillo’

   See section ‘The DESCRIPTION file’ in the ‘Writing R Extensions’
   manual.

   See
     ‘/Users/andybega/Work/BayesMFSurv.Rcheck/00check.log’
   for details.

── R CMD check results ────────────────────────────────── BayesMFSurv 0.2.0 ────
Duration: 3.9s

> checking package dependencies ... ERROR
  Namespace dependency not required: ‘RcppArmadillo’

  See section ‘The DESCRIPTION file’ in the ‘Writing R Extensions’
  manual.

1 error x | 0 warnings ✓ | 0 notes ✓
Error: R CMD check found ERRORs
Execution halted

Exited with status 1.
```

Adding RcppArmadillo to imports in the DESCRIPTION fixes this and R check runs without errors, warnings, notes.
