### 2020-03-20 Nicolas-Schmidt/BayesMFSurv#2 @andybega
Both authors and authors@R in DESCRIPTION
Both the the Author and Authors@R fields in the DESCRIPTION are specified, and the entries don't match. Maybe just keep Authors@R?

https://github.com/openjournals/joss-reviews/issues/2164
