### 2020-03-20 Nicolas-Schmidt/BayesMFSurv#3 @andybega
Example in README doesn't work, data is not loaded
The README examples don't run as they are. Needs a `data(RBS)`, around line 90-91 in README.Rmd.
