### 2020-03-20 #2164 @andybega
Hi @Nicolas-Schmidt and @Sergio-Bejar,

I've opened a couple of issues in your repo; below are the ones I think are needed for the review/checklist. The others are suggestions that I think would make the package better, but I don't think it's needed for the review.

- Duplicate Authors fields in DESCRIPTION https://github.com/Nicolas-Schmidt/BayesMFSurv/issues/2
- Missing `data()` call in README example https://github.com/Nicolas-Schmidt/BayesMFSurv/issues/3
- The package version referenced in the paper is 0.2.0. The version on CRAN is 0.1.0. Are you planning to submit 0.2.0 to CRAN before this review is over? Ok with me if not, but then I would ask you to fix two minor issues related to documentation and R check in https://github.com/Nicolas-Schmidt/BayesMFSurv/issues/1

One minor thing, for the paper, I think it would be helpful in the first paragraph to have 2 or so examples where mis-classified failures might be a problem. E.g. the civil war example from the PA paper (we don't really know that well how many people die in remote conflicts, so some civil wars might be coded as ended when they are not), and then maybe one other non-poli-sci example? Or I guess the democracy one in a pickle.

### Testing

I guess ideally you'd have some combination of `testthat` unit tests and some kind of CI like travis. Since the package is pretty small and you have examples in the docs for the main functions that would show up in R check if they broke, maybe just having CI would meet this?

When I cloned the repo and built the package from source, I found a missing dependency and some documentation that had to be updated (including changes to NAMESPACE). That would help you catch issues like this.

(see e.g. `?usethis::use_travis(ext = "com")` for setup)
