+++ README.utf8.md
@@ -4,14 +4,7 @@ output: github_document
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 
-```{r, include = FALSE}
-knitr::opts_chunk$set(
-  collapse = TRUE,
-  comment = "#>",
-  fig.path = "man/figures/README-",
-  out.width = "100%"
-)
-```
+
 
 ## The `BayesMFSurv` package
 
@@ -81,7 +74,8 @@ The data used to estimate the following examples come from Reenock, Bernhard and
 
 
 #### Misclassified-Failure
-```{r example, eval = FALSE}
+
+```r
 
 # Baseline Bayesian misclassified failure (MF) model. 
 # Misclassification stage only includes the intercept while the survival stage 
@@ -105,22 +99,65 @@ model1 <- mfsurv(Y ~ X | C ~ Z1, Y0 = Y0,
                  form = "Weibull",
                  na.action = 'na.omit')
 
-```{r, echo = FALSE}
-library(BayesMFSurv)
-model1 <- rio::import(here::here("data-raw", "model1_output.rds"))
 ```
 
-```{r}
+
+
+```r
 stats(model1)
+#> $DIC
+#> [1] 466.1642
+#> 
+#> $Loglik
+#> [1] 1243.269
 
 summary(model1, parameter = c("betas"))
+#> 
+#> Iterations = 1:900
+#> Thinning interval = 1 
+#> Number of chains = 1 
+#> Sample size per chain = 900 
+#> 
+#> 1. Empirical mean and standard deviation for each variable,
+#>    plus standard error of the mean:
+#> 
+#>                 Mean     SD Naive SE Time-series SE
+#> X.intercept  1.59992 2.5488  0.08496        0.11147
+#> X1           1.56001 1.6683  0.05561        0.08020
+#> Xcalinv      0.16505 2.5411  0.08470        0.19144
+#> Xlnlevel    -1.12098 1.4762  0.04921        0.09873
+#> Xcalileve   -0.03761 2.4731  0.08244        0.15828
+#> Xnecon      -2.13090 2.4107  0.08036        0.07463
+#> Xpresi       0.23041 1.9100  0.06367        0.05790
+#> Xtag        -0.01230 0.7878  0.02626        0.03822
+#> Xrel         1.45198 1.7404  0.05801        0.04774
+#> Xethn        1.01890 1.3000  0.04333        0.03564
+#> Xprevdem     0.13070 1.2525  0.04175        0.11232
+#> Xopenc      -0.19741 1.7926  0.05975        0.18939
+#> 
+#> 2. Quantiles for each variable:
+#> 
+#>                 2.5%      25%      50%     75%     97.5%
+#> X.intercept -1.13533  0.69903  1.52827  2.5172  4.865523
+#> X1          -1.40643  0.62558  1.55816  2.4422  4.804547
+#> Xcalinv     -3.85120 -1.13823  0.00846  1.2496  4.676094
+#> Xlnlevel    -1.42372 -1.12625 -1.01540 -0.8972 -0.676814
+#> Xcalileve   -3.85666 -1.07941  0.04700  1.2389  4.015589
+#> Xnecon      -6.11097 -3.09541 -2.00216 -0.9263  1.233515
+#> Xpresi      -0.38909  0.05153  0.26004  0.4705  0.945025
+#> Xtag        -0.25624 -0.06322  0.03359  0.1123  0.255185
+#> Xrel         0.07307  0.93358  1.38643  1.8464  2.834912
[…]
