+++ NAMESPACE
@@ -4,6 +4,8 @@ S3method(summary,mcmcsurv)
 S3method(summary,mfsurv)
 export(mcmcsurv)
 export(mfsurv)
+export(mfsurv.stats)
+export(mfsurv.summary)
 export(stats)
 import(grDevices)
 import(graphics)
