+++ README.knit.md
@@ -1,240 +1,0 @@
----
-output: github_document
----
-
-<!-- README.md is generated from README.Rmd. Please edit that file -->
-
-
-
-## The `BayesMFSurv` package
-
-_Minnie M. Joo, Nicolas Schmidt, Sergio Bejar, Bumba Mukherjee, Vineeta Yadav_
-
-<!-- badges: start -->
-[![CRAN\_Status\_Badge](https://www.r-pkg.org/badges/version/BayesMFSurv)](https://cran.r-project.org/package=BayesMFSurv)
-[![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
-[![CRAN RStudio mirrordownloads](https://cranlogs.r-pkg.org/badges/grand-total/BayesMFSurv?color=blue)](https://www.r-pkg.org/pkg/BayesMFSurv)
-[![CRAN RStudio mirrordownloads](https://cranlogs.r-pkg.org/badges/BayesMFSurv?color=blue)](https://www.r-pkg.org/pkg/BayesMFSurv)
-<!-- badges: end -->
-
-
-### Description
-
-Contains a split population survival estimator that models the misclassification probability of failure versus right-censored events. The split population survival estimator is described in Bagozzi et al. (2019) <doi:10.1017/pan.2019.6>.
-
-### Installation
-
-You can install the released version (`0.1.0`) of `BayesMFSurv` from [CRAN](https://cran.r-project.org/) with:
-
-``` r
-install.packages("BayesMFSurv")
-```
-
-And the development version (`0.2.0`) from GitHub with:
-
-``` r
-if (!require("remotes")) install.packages("remotes")
-remotes::install_github("Nicolas-Schmidt/BayesMFSurv")
-```
-
-## Functions
-
-
-| Function |Description |
-| ----------------- | -------------------------------------------------------------- | 
-|`mfsurv`|fits a parametric Bayesian MF model via Markov Chain Monte Carlo (MCMC) to estimate the probability of misclassification in the first stage and the hazard in the second stage. Slice sampling is employed to draw the posterior sample of the model's split and survival stage parameters.|
-|`mcmcsurv` | estimates a Bayesian Exponential or Weibull survival model via Markov Chain Monte Carlo (MCMC). Slice samplig is employed to draw the posterior sample of the model's survival stage parameters.|
-|`stats`|a function to calculate the deviance information criterion (DIC) and the log-likelihood for fitted model objects of class mfsurv or mcmcsurv.|
-|`summary`|returns a summary of a mfsurv or mcmcsurv object via `coda::summary.mcmc`.|
-
-
-
-
-
-### Example
-
-
-The data used to estimate the following examples come from Reenock, Bernhard and Sobek (2007) -DOI: 10.111/j.1468-2478.2007.00469.x-.  The RBS (2007) dataset uses continuous-time event history techniques to code episodes of democratic breakdown in all democracies from 1961 to 1995.  In addition, it provides data on a number of economic and political variables. 
-
-|Variable| Description|
-| ----------------- | -------------------------------------------------------------- | 
[…]
