+++ paper.md
@@ -38,7 +38,7 @@ affiliations:
 
 # Summary
 
-Social Scientists and Biostatisticians often employ conventional parametric survival or mixture cure models (e.g. Weibull, Exponential) to analyze outcome variables in survival data that focus on the time until an event occurred or "failed" [@box1999modeling; @maller1996survival; @lee17]. An important assumption underlying these models is that researchers record the date -year, month or day- in which an event or observation of interest failed (i.e. "terminated") _accurately_. Yet events that are recorded as having failed at a given point of time can be inaccurately measured [@clark2003survival; @schober2018survival; @bagozzi2019bayesian]. Inaccurate measurement of this sort leads to a subset of _misclassified failure_ cases in survival data in which some subjects are recorded as having failed or experienced the event of interest even though they in actuality "live on" past their recorded-failure point. Conventional survival and cure models yield biased parameter estimates in survival datasets that include latent misclassified failure cases as they do not statistically account for such cases [@schober2018survival; @bagozzi2019bayesian].
+Social Scientists and Biostatisticians often employ conventional parametric survival or mixture cure models (e.g. Weibull, Exponential) to analyze outcome variables in survival data that focus on the time until an event occurred or "failed" [@box1999modeling; @maller1996survival; @lee17]. An important assumption underlying these models is that researchers record the date -year, month or day- in which an event or observation of interest failed (i.e. "terminated") _accurately_. Yet events that are recorded as having failed at a given point of time can be inaccurately measured [@clark2003survival; @schober2018survival; @bagozzi2019bayesian]. Inaccurate measurement of this sort leads to a subset of _misclassified failure_ cases in survival data in which some subjects are recorded as having failed or experienced the event of interest even though they in actuality "live on" past their recorded-failure point. In political science, for instance, scholars that study the determinants of civil war duration often record civil war end dates ("failures") using overly conservative tresholds, especially for lower-intensity conflicts in remote or poor information environments that persist indefinitely with very little fighting [@buhaugh2009geography; @thyne2012information]. Moreover, Conventional survival and cure models yield biased parameter estimates in survival datasets that include latent misclassified failure cases as they do not statistically account for such cases [@schober2018survival; @bagozzi2019bayesian].
 
 As described below, [@bagozzi2019bayesian] develop a Bayesian parametric Misclassified Failure (MF) survival model which -similar to standard cure models [@Zhouetal-spBayes; @box1999modeling; @maller1996survival]- consists of a system of two equations that addresses misclassified event failures in survival data. `BayesMFSurv` [@BayesMFSurv-v1] is an R package [@r-core] that contains functions and computationally intensive routines in `C++` to fit the parametric Weibull and Exponential (i) survival model and (ii) Misclassified Failure survival model via Bayesian Markov Chain Monte Carlo (MCMC) methods using slice-sampling [@neal2003slice; @bagozzi2019bayesian].
 
