+++ paper.bib
@@ -1,3 +1,15 @@
+@article{buhaug2009geography,
+  title    = {Geography, rebel capability, and the duration of civil conflict},
+  author   = {Buhaug, Halvard and Gates, Scott and Lujala, P{\"a}ivi},
+  journal  = {Journal of Conflict Resolution},
+  volume   = {53},
+  number   = {4},
+  pages    = {544--569},
+  year     = {2009},
+  publisher= {Sage Publications Sage CA: Los Angeles, CA},
+  doi      = {10.1177/0022002709336457}
+}  
+
 @article{clark2003survival,
   title     = {Survival analysis part I: basic concepts and first analyses},
   author    = {Clark, Taane G and Bradburn, Michael J and Love, Sharon B and Altman, Douglas G},
