+++ paper.md
@@ -38,7 +38,11 @@ affiliations:
 
 # Summary
 
-Social Scientists and Biostatisticians often employ conventional parametric survival or mixture cure models (e.g. Weibull, Exponential) to analyze outcome variables in survival data that focus on the time until an event occurred or "failed" [@box1999modeling; @maller1996survival; @lee17]. An important assumption underlying these models is that researchers record the date -year, month or day- in which an event or observation of interest failed (i.e. "terminated") _accurately_. Yet events that are recorded as having failed at a given point of time can be inaccurately measured [@clark2003survival; @schober2018survival; @bagozzi2019bayesian]. Inaccurate measurement of this sort leads to a subset of _misclassified failure_ cases in survival data in which some subjects are recorded as having failed or experienced the event of interest even though they in actuality "live on" past their recorded-failure point. In political science, for instance, scholars that study the determinants of civil war duration often record civil war end dates ("failures") using overly conservative tresholds, especially for lower-intensity conflicts in remote or poor information environments that persist indefinitely with very little fighting [@buhaugh2009geography; @thyne2012information]. [@schober2018survival] document a similar issue in medical research analyzing the determinants of a well-defined event such as patient mortality. Conventional survival and cure models yield biased parameter estimates in survival datasets that include latent misclassified failure cases as they do not statistically account for such cases [@schober2018survival; @bagozzi2019bayesian].
+Social Scientists and Biostatisticians often employ conventional parametric survival or mixture cure models (e.g. Weibull, Exponential) to analyze outcome variables in survival data that focus on the time until an event occurred or "failed" [@box1999modeling; @maller1996survival; @lee17]. An important assumption underlying these models is that researchers record the date -year, month or day- in which an event or observation of interest failed (i.e. "terminated") _accurately_. Yet events that are recorded as having failed at a given point of time can be inaccurately measured [@clark2003survival; @schober2018survival; @bagozzi2019bayesian]. Inaccurate measurement of this sort leads to a subset of _misclassified failure_ cases in survival data in which some subjects are recorded as having failed or experienced the event of interest even though they in actuality "live on" past their recorded-failure point. 
+
+There are several scenarios where a subset of recorded failure events may persist beyond their recorded failure time, leading to misclassiﬁcation in event failures. For example, political scientists who analyze the duration of civil wars fought between rebel groups and governments often record civil war end dates (“failures”) for specific conﬂicts based upon 24-month spells with fewer than 25 battle-deaths per year (Balch-Lindsay and Enterline 2000; @thyne2012information]. This 25-battle death threshold is prone to measurement error, especially for lower-intensity civil wars in poor information environments that persist beyond their recorded end date. Other examples include the study of the duration of ancient civilizations [@cioffi1999evolution] and the time taken to detect cancer [@schober2018survival]. In both these latter examples, researchers typically do not have data on the precise time-point of a given failure due to the sands of time or because of inaccurate information. This leads them to—similar to the civil conflict example—underestimate the duration of particularly misclassified event failure cases.
+
+Conventional survival and cure models yield biased parameter estimates in survival datasets that include latent misclassified failure cases as they do not statistically account for such cases [@schober2018survival; @bagozzi2019bayesian].
 
[…]
