+++ paper.bib
@@ -1,14 +1,14 @@
-@article{buhaug2009geography,
-  title    = {Geography, rebel capability, and the duration of civil conflict},
-  author   = {Buhaug, Halvard and Gates, Scott and Lujala, P{\"a}ivi},
-  journal  = {Journal of Conflict Resolution},
-  volume   = {53},
+@article{balch2000killing,
+  title    = {Killing time: The world politics of civil war duration, 1820--1992},
+  author   = {Balch-Lindsay, Dylan and Enterline, Andrew J},
+  journal  = {International Studies Quarterly},
+  volume   = {44},
   number   = {4},
-  pages    = {544--569},
-  year     = {2009},
-  publisher= {Sage Publications Sage CA: Los Angeles, CA},
-  doi      = {10.1177/0022002709336457}
-}  
+  pages    = {615--642},
+  year     = {2000},
+  publisher={Blackwell Publishers, Inc. Boston, USA and Oxford, UK},
+  doi      = {10.1111/0020-8833.00174}
+}
 
 @article{clark2003survival,
   title     = {Survival analysis part I: basic concepts and first analyses},
