+++ README.md
@@ -86,69 +86,68 @@ variables.
 library(BayesMFSurv)
 
 set.seed(95)
-RBS <- data(RBS)
+data(RBS)
+RBS <- na.omit(RBS)
 Y   <- RBS$Y
 X   <- as.matrix(cbind(1, RBS[,1:10]))
 C   <- RBS$C
 Z1  <- cbind(rep(1,nrow(RBS)))
 Y0  <- RBS$Y0
 model1 <- mfsurv(Y ~ X | C ~ Z1, Y0 = Y0,
-                 N = 100000,
-                 burn = 10000,
-                 thin = 100,
-                 w = c(0.5, .5, .5),
+                 N = 100,
+                 burn = 50,
+                 thin = 5,
+                 w = c(0.5, 0.5, 0.5),
                  m = 20,
-                 form = "Weibull",
-                 na.action = 'na.omit')
-```
+                 form = 'Weibull')
 
-``` r
+
 stats(model1)
 #> $DIC
-#> [1] 466.1642
+#> [1] 256.8576
 #> 
 #> $Loglik
-#> [1] 1243.269
+#> [1] 478.8454
 
 summary(model1, parameter = c("betas"))
 #> 
-#> Iterations = 1:900
+#> Iterations = 1:10
 #> Thinning interval = 1 
 #> Number of chains = 1 
-#> Sample size per chain = 900 
+#> Sample size per chain = 10 
 #> 
 #> 1. Empirical mean and standard deviation for each variable,
 #>    plus standard error of the mean:
 #> 
-#>                 Mean     SD Naive SE Time-series SE
-#> X.intercept  1.59992 2.5488  0.08496        0.11147
-#> X1           1.56001 1.6683  0.05561        0.08020
-#> Xcalinv      0.16505 2.5411  0.08470        0.19144
-#> Xlnlevel    -1.12098 1.4762  0.04921        0.09873
-#> Xcalileve   -0.03761 2.4731  0.08244        0.15828
-#> Xnecon      -2.13090 2.4107  0.08036        0.07463
-#> Xpresi       0.23041 1.9100  0.06367        0.05790
-#> Xtag        -0.01230 0.7878  0.02626        0.03822
-#> Xrel         1.45198 1.7404  0.05801        0.04774
-#> Xethn        1.01890 1.3000  0.04333        0.03564
-#> Xprevdem     0.13070 1.2525  0.04175        0.11232
-#> Xopenc      -0.19741 1.7926  0.05975        0.18939
+#>                 Mean    SD Naive SE Time-series SE
+#> X.intercept  -5.0477 3.384   1.0700         1.0700
+#> X1          -10.6678 3.797   1.2006         1.2006
+#> Xcalinv      12.8464 7.331   2.3182         6.9599
+#> Xlnlevel     -3.2283 3.595   1.1367         1.1367
+#> Xcalileve     0.9831 2.667   0.8435         0.8369
+#> Xnecon       -8.9798 4.207   1.3302         2.9719
+#> Xpresi       -2.4573 3.032   0.9588         0.9588
+#> Xtag         -3.7535 2.600   0.8222         0.8222
+#> Xrel         -8.1554 3.174   1.0037         2.0935
+#> Xethn         2.2446 3.124   0.9878         1.9027
+#> Xprevdem     13.8614 2.904   0.9182         0.9182
+#> Xopenc       -2.4537 1.367   0.4324         0.4324
 #> 
 #> 2. Quantiles for each variable:
 #> 
[…]
