+++ man/mfsurv.stats.Rd
@@ -1,0 +1,20 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/deprecated.R
+\name{mfsurv.stats}
+\alias{mfsurv.stats}
+\title{mfsurv.stats}
+\usage{
+mfsurv.stats(object)
+}
+\arguments{
+\item{object}{an object of class \code{mfsurv}, the output of \code{mfsurv()}.}
+}
+\value{
+list.
+}
+\description{
+A function to calculate the deviance information criterion (DIC) for fitted model objects of class \code{mfsurv}
+for which a log-likelihood can be obtained, according to the formula \emph{DIC = -2 * (L - P)},
+where \emph{L} is the log likelihood of the data given the posterior means of the parameter and
+\emph{P} is the  estimate of the effective number of parameters in the model.
+}

[…]
