### 2020-03-23 #2164 @alletsee
Hi @Nicolas-Schmidt and @Sergio-Bejar,

I saw that you have included @andybega's suggestion of providing examples where misclassified failures are problematic; I second his suggestion and think that your edit addresses this well. (I think you have not yet generated a new article proof.)

Apart from this, I have only a few minor edit suggestions for the paper that are mainly about form:

- You introduce the "Misclassified Failure (MF)" abbreviation multiple times; introducing it once should be sufficient.
-  There are a few instances with in-text citation where the full citation appears in parentheses while the author names should be part of the main sentence: "As described below, (Bagozzi et al., 2019) develop a Bayesian parametric" in the second paragraph of the summary and "using a slice-sampling algorithm described in (Bagozzi et al., 2019)." in the first paragraph of the "BayesMFSurv R Package" chapter. It would be great if you could use proper in-text citation in those examples ([see here](https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html#citation_syntax).
- In the last paragraph of the "Motivation, Description, Applications" chapter, there is a citation in the middle of the sentence that looks like a formatting error: "our BayesMFSurv R package incorporates various functions listed below that fit (Bagozzi et al., 2019) parametric Misclassified Failure (MF) survival model via Bayesian MCMC methods." Did you maybe mean to have this at the end of the sentence
- There were sometimes minor inconsistencies in the capitalization of the journal names in the bibliography, for instance "International Studies Quarterly" vs. "British journal of cancer".

(As a quick update, I would still need to review the package functionality.)
