+++ README.Rmd
@@ -87,6 +87,8 @@ The data used to estimate the following examples come from Reenock, Bernhard and
 # Misclassification stage only includes the intercept while the survival stage 
 # includes all covariates described above.  
 
+# Examples that use a larger N can be found here:  https://github.com/Nicolas-Schmidt/BayesMFSurv/tree/master/data-raw
+
 library(BayesMFSurv)
 
 set.seed(95)
