+++ README.md
@@ -83,6 +83,8 @@ variables.
 # Misclassification stage only includes the intercept while the survival stage 
 # includes all covariates described above.  
 
+# Examples that use a larger N can be found here:  https://github.com/Nicolas-Schmidt/BayesMFSurv/tree/master/data-raw
+
 library(BayesMFSurv)
 
 set.seed(95)
