+++ paper.md
@@ -42,17 +42,17 @@ Social Scientists and Biostatisticians often employ conventional parametric surv
 
 There are several scenarios where a subset of recorded failure events may persist beyond their recorded failure time, leading to misclassiﬁcation in event failures. For example, political scientists who analyze the duration of civil wars fought between rebel groups and governments often record civil war end dates (“failures”) for specific conﬂicts based upon 24-month spells with fewer than 25 battle-deaths per year [@balch2000killing; @thyne2012information]. This 25-battle death threshold is prone to measurement error, especially for lower-intensity civil wars in poor information environments that persist beyond their recorded end date. Other examples include the study of the duration of ancient civilizations [@cioffi1999evolution] and the time taken to detect cancer [@schober2018survival]. In both these latter examples, researchers typically do not have data on the precise time-point of a given failure due to the sands of time or because of inaccurate information. This leads them to—similar to the civil conflict example—underestimate the duration of particularly misclassified event failure cases.
 
-Since these underestimates of duration are non-random, bias will arise in survival estimates of the phenomena mentioned above when using conventional survival models. Hence, the main motivation for developing the Bayesian Misclassiﬁed Failure (MF) split population survival model is to resolve methodological challenges resulting from misclassiﬁed event failures by accounting for the possibility that some failure events survive beyond their recorded failure time. The development of the Bayesian MF model is also driven by the fact that it permits researchers to identify when the end-date of observations in survival data is misclassiﬁed therein providing substantive insights into this process. Finally, there does not exist an R package that extracts posterior distribution of estimates from parametric cure (split-population) models, including the MF model, using Bayesian Markov Chain Monte Carlo (MCMC) methods. `BayesMFSurv` [@BayesMFSurv-v1] is an R package [@r-core] that contains functions and computationally intensive routines in `C++` to fit the parametric Weibull and Exponential (i) survival model and (ii) Misclassified Failure survival model via Bayesian MCMC methods using slice-sampling [@neal2003slice; @bagozzi2019bayesian].
+Since these underestimates of duration are non-random, bias will arise in survival estimates of the phenomena mentioned above when using conventional survival models. Hence, the main motivation for developing the Bayesian Misclassiﬁed Failure (MF hereafter) split population survival model is to resolve methodological challenges resulting from misclassiﬁed event failures by accounting for the possibility that some failure events survive beyond their recorded failure time. The development of the Bayesian MF model is also driven by the fact that it permits researchers to identify when the end-date of observations in survival data is misclassiﬁed therein providing substantive insights into this process. Finally, there does not exist an R package that extracts posterior distribution of estimates from parametric cure (split-population) models, including the MF model, using Bayesian Markov Chain Monte Carlo (MCMC) methods. `BayesMFSurv` [@BayesMFSurv-v1] is an R package [@r-core] that contains functions and computationally intensive routines in `C++` to fit the parametric Weibull and Exponential (i) survival model and (ii) Misclassified Failure survival model via Bayesian MCMC methods using slice-sampling [@neal2003slice; @bagozzi2019bayesian].
 
 # Motivation, Description, Applications
 
[…]
