+++ paper.bib
@@ -13,7 +13,7 @@
 @article{clark2003survival,
   title     = {Survival analysis part I: basic concepts and first analyses},
   author    = {Clark, Taane G and Bradburn, Michael J and Love, Sharon B and Altman, Douglas G},
-  journal   = {British journal of cancer},
+  journal   = {British Journal of Cancer},
   volume    = {89},
   number    = {2},
   pages     = {232--238},
@@ -36,7 +36,7 @@
 @article{price2001modelling,
   title     = {Modelling survival data with a cured fraction using frailty models},
   author    = {Price, Dionne L and Manatunga, Amita K},
-  journal   = {Statistics in medicine},
+  journal   = {Statistics in Medicine},
   volume    = {20},
   number    = {9-10},
   pages     = {1515--1527},
@@ -55,7 +55,7 @@
 @article{neal2003slice,
   title     = {Slice sampling},
   author    = {Neal, Radford M},
-  journal   = {Annals of statistics},
+  journal   = {Annals of Statistics},
   pages     = {705--741},
   year      = {2003},
   publisher = {JSTOR},
@@ -77,7 +77,7 @@
 @article{schober2018survival,
   title     = {Survival analysis and interpretation of time-to-event data: the tortoise and the hare},
   author    = {Schober, Patrick and Vetter, Thomas R},
-  journal   = {Anesthesia and analgesia},
+  journal   = {Anesthesia and Analgesia},
   volume    = {127},
   number    = {3},
   pages     = {792},
@@ -119,7 +119,7 @@
 }
 
 @book{maller1996survival,
-  title     = {Survival analysis with long-term survivors},
+  title     = {Survival Analysis with Long-Term Survivors},
   author    = {Maller, Ross A and Zhou, Xian},
   year      = {1996},
   publisher = {John Wiley \& Sons},
@@ -148,9 +148,10 @@
   title     = {{spBayesSurv}: Fitting Bayesian Spatial Survival Models Using {R}},
   author    = {Haiming Zhou and Timothy Hanson and Jiajia Zhang},
   journal   = {Journal of Statistical Software},
-  year      = {2018},
-  volume    = {accept minor},
-  url       = {http://arxiv.org/abs/1705.04584},
+  volume    = {92},
+  number    = {9},
+  pages     = {1--33},
+  year      = {2020},
   doi       = {10.18637/jss.v092.i09}
 }
 
@@ -187,7 +188,7 @@
 }
 
 @article{cioffi1999evolution,
-  title     = {Evolution of Maya polities in the ancient Mesoamerican system},
+  title     = {Evolution of Maya Polities in the Ancient Mesoamerican System},
   author    = {Cioffi-Revilla, Claudio and Landman, Todd},
   journal   = {International Studies Quarterly},
   volume    = {43},
