+++ paper.md
@@ -48,7 +48,7 @@ Since these underestimates of duration are non-random, bias will arise in surviv
 
 Numerous R packages offer functionalities to estimate conventional parametric and semi-parametric survival models via maximum likelihood estimation (MLE) or Bayesian MCMC methods [@OIsurv; @Zhouetal-spBayes; @survival-package; @wangetal-dynsurv]. Other R packages focus on estimation of parametric or semi-parametric cure survival models using MLE [@smcure-cai; @beger2017splitting; @rcure-r; @amdahl-flex]. To our knowledge, there does not exist an R package that fits parametric mixture cure models, including the MF survival model, via _BayesianMCMC_ (e.g. slice sampling) methods which offer a powerful yet flexible tool for estimating such models. Further, extant R packages that use Bayesian inference for survival analyses only focus on standard survival models that do not take into account latent misclassified failure events in survival data. Because misclassified failure events are right-censored events, there is a non-zero probability that these misclassified cases persisted beyond their recorded failure time. Failing to account for misclassified failure events in survival data that results from estimating standard survival or cure models will lead researchers to underestimate the duration of time of these events.
 
-Since the underestimates of duration are non-random, bias will arise in survival estimates of these phenomena when researchers use standard survival or cure models. To address this misclassified failure challenge in survival data, our `BayesMFSurv` R package incorporates various functions listed below that fit @bagozzi2019bayesian's parametric MF survival model via Bayesian MCMC methods. This model estimates a system of two equations to account for the possibility that some unknown subset of failure events actually "lived on" beyond their recorded failure time. The first is a "splitting" equation that estimates the probability of a case being a misclassified failure, with or without covariates. The second equation is a standard parametric survival model, whose relevant failure and survival probabilities are estimated conditional on a case _not_ being a misclassified failure. These features of the model in `BayesMFSurv` account for a heterogeneous mixture of failure cases in survival data and address the non-random underestimates of duration for misclassified failure events. `BayesMFSurv` also incorporates time-varying covariates that are common in panel survival datasets and this model can be applied to the following survival datasets where misclassified failure cases are prevalent: civil war termination that determines civil war duration [@thyne2012information], time taken to detect onset of cancer [@schober2018survival], and collapse (and thus duration) of ancient civilizations or political regimes [@cioffi1999evolution; @reenock2007regressive].
+Since the underestimates of duration are non-random, bias will arise in survival estimates of these phenomena when researchers use standard survival or cure models. To address this misclassified failure challenge in survival data, our `BayesMFSurv` R package incorporates various functions listed below that fit @bagozzi2019bayesian's parametric MF survival model via _BayesianMCMC_ methods. This model estimates a system of two equations to account for the possibility that some unknown subset of failure events actually "lived on" beyond their recorded failure time. The first is a "splitting" equation that estimates the probability of a case being a misclassified failure, with or without covariates. The second equation is a standard parametric survival model, whose relevant failure and survival probabilities are estimated conditional on a case _not_ being a misclassified failure. These features of the model in `BayesMFSurv` account for a heterogeneous mixture of failure cases in survival data and address the non-random underestimates of duration for misclassified failure events. `BayesMFSurv` also incorporates time-varying covariates that are common in panel survival datasets and this model can be applied to the following survival datasets where misclassified failure cases are prevalent: civil war termination that determines civil war duration [@thyne2012information], time taken to detect onset of cancer [@schober2018survival], and collapse (and thus duration) of ancient civilizations or political regimes [@cioffi1999evolution; @reenock2007regressive].
 
 # BayesMFSurv R Package
 
