+++ README.Rmd
@@ -87,7 +87,7 @@ The data used to estimate the following examples come from Reenock, Bernhard and
 # Misclassification stage only includes the intercept while the survival stage 
 # includes all covariates described above.  
 
-# Examples that use a larger N can be found here:  https://github.com/Nicolas-Schmidt/BayesMFSurv/tree/master/data-raw
+# Example that use a larger N can be found here:  https://github.com/Nicolas-Schmidt/BayesMFSurv/tree/master/data-raw
 
 library(BayesMFSurv)
 
@@ -115,7 +115,9 @@ summary(model1, parameter = c("betas"))
 
 #### Non Misclassified-Failure
 
+`mcmcsurv` estimates a Bayesian equivalent of standard survival models (i.e. Exponential or Weibull).   
 
+#Example with N=15000 can be found here:  https://github.com/Nicolas-Schmidt/BayesMFSurv/tree/master/data-raw
 
 ```{r}
 set.seed(95)
