+++ README.Rmd
@@ -56,8 +56,6 @@ remotes::install_github("Nicolas-Schmidt/BayesMFSurv")
 
 
 
-
-
 ### Example
 
 
@@ -81,13 +79,18 @@ The data used to estimate the following examples come from Reenock, Bernhard and
 
 
 #### Misclassified-Failure
+
+`mfsurv` estimated the probability of misclassification failure in the first (split) stage and hazard in the second (survival) stage. 
+
+`mfsurv` should be used when user suspects that data of survival cases could be right-sensored (i.e. when there is a probability that failure events are misclassified). 
+
+Example with N = 100,000 is here: https://github.com/Nicolas-Schmidt/BayesMFSurv/tree/master/data-raw
+
 ```{r}
 
 # Baseline Bayesian misclassified failure (MF) model. 
 # Misclassification stage only includes the intercept while the survival stage 
 # includes all covariates described above.  
-
-# Example that use a larger N can be found here:  https://github.com/Nicolas-Schmidt/BayesMFSurv/tree/master/data-raw
 
 library(BayesMFSurv)
 
