+++ paper.bib
@@ -48,7 +48,7 @@
 @inproceedings{box1999modeling,
   title     = {Modeling heterogeneity in duration models},
   author    = {Box-Steffensmeier, Janet M and Zorn, Christopher},
-  booktitle = {Summer Meeting of the Political Methodology Society, July15-17},
+  booktitle = {Summer Meeting of the Political Methodology Society, July 15-17},
   year      = {1999}
 }
 
