### 2020-03-23 Nicolas-Schmidt/BayesMFSurv#1 @Sergio-Bejar
Hi Andy,

Thanks for the time you spent reviewing our submission.   Below, please find our response to this point.

R: This is a good point.  We had a similar issue when we submitted our package to CRAN R.  The problem itself is not necessarily with the dependency of 'RccpArmadillo', but with the default compilation of the 'NAMESPACE' file that is generated when one executes devtools::documents().  Specifically, the aforementioned function generates importFrom[(RccpArmadillo & importFrom(RccpsourceCpp)] in the 'NAMESPACE' file when it should only import 'Rccp' in 'evalCpp' and not ''RccpArmadillo'.

Therefore, to prevent issues/errors during the R CMD check when using devtools it is necessary to be specific about the files one wants to compile.  That is, we should compile the files with extension rd but not the 'NAMESPACE' file.    To compile the package  in a local station, we suggest using: devtools::documents(roclets = 'rd').
