### 2020-03-23 Nicolas-Schmidt/BayesMFSurv#2 @Sergio-Bejar
We updated the description file in our repo.  Thanks for catching this mistake!
