### 2020-03-23 Nicolas-Schmidt/BayesMFSurv#3 @Sergio-Bejar
As suggested, we added data(RBS).  Please see line 98 of the README.Rmd file.
