### 2020-03-23 #2164 @Sergio-Bejar
Hi @andybega and @alletsee,

We have addressed all the points raised by you in your previous messages.   I already responded in detail to all the issues raised by @andybega in our Github repo (please see those responses here: [https://github.com/Nicolas-Schmidt/BayesMFSurv/issues](url).  Below, I respond to the points raised in this page.

@andybega:

- We addressed points 1 and 2 as suggested by you.
[https://github.com/Nicolas-Schmidt/BayesMFSurv/issues/2](url)
[https://github.com/Nicolas-Schmidt/BayesMFSurv/issues/3](url)

- We also responded to issue #1 raised by you. We believe that our response addresses the issues related with (i) documentation and (ii) R CMD checks that you found [https://github.com/Nicolas-Schmidt/BayesMFSurv/issues/1](url).   We believe that the issues you found prompted your comment related to using travis CI (see: **Testing**).  If this is not the case, please let us know. We will be happy to incorporate travis CI into the package.

- Version 0.2.0 of our package is ready to be submitted to CRAN R, we wanted to first hear comments/suggestions from our reviewers in order to improve the package.  We plan to submit version 0.2.0 to CRAN R as soon as this review is over.

- Following your suggestion, the revised version of our manuscript now includes three examples that illustrate when mis-classified failures might be a problem -see paragraph 2-.  Further, we also explain how such mis-classification may affect the results one obtains when using standard survival models.

@alletsee:

- We carefully edited the manuscript to address all your comments.  Specifically, we tried hard to avoid being redundant when mentioning the name of the model.  We also edited the citations in the the sections of the paper, and the bib file containing the references -i.e. being consistent with the capitalization of journal names, and adding the correct reference for Zhou et al (2020).

Thanks again for all your comments and suggestions!
