+++ README.Rmd
@@ -85,7 +85,7 @@ The data used to estimate the following examples come from Reenock, Bernhard and
 
 `mfsurv` should be used when user suspects that data of survival cases could be right-sensored (i.e. when there is a probability that failure events are misclassified). 
 
-Example with N = 100,000 is [here](https://github.com/Nicolas-Schmidt/BayesMFSurv/tree/master/data-raw).
+Example with N = 100000 is [here](https://github.com/Nicolas-Schmidt/BayesMFSurv/tree/master/data-raw).
 
 ```{r}
 
@@ -122,7 +122,7 @@ summary(model1, parameter = c("betas"))
 
 `mcmcsurv` estimates a Bayesian equivalent of standard survival models (i.e. Exponential or Weibull).   
 
-#Example with N=15000 can be found here:  https://github.com/Nicolas-Schmidt/BayesMFSurv/tree/master/data-raw
+Example with N = 15000 can be found [here]( https://github.com/Nicolas-Schmidt/BayesMFSurv/tree/master/data-raw).
 
 ```{r}
 set.seed(95)
