### 2020-03-24 Nicolas-Schmidt/BayesMFSurv#1 @andybega
Hi Sergio!
`mfsurv.R` has directives at the top to import `RcppArmadillo` and `Rcpp::sourceCpp`:

```r
#' @import RcppArmadillo
#' @importFrom Rcpp sourceCpp
```

Since those are not needed (yes?), what about just taking them out? That will fix the problem with using plain `devtools::document()`, i.e. including NAMESPACE generation, and also fixes the R check error.
