### 2020-03-24 #2164 @andybega
Hi @Nicolas-Schmidt and @Sergio-Bejar,

thanks for addressing the points raised. Two items I would like to keep open:

- Re the [1st issue, R check error](https://github.com/Nicolas-Schmidt/BayesMFSurv/issues/1) (summarizing for the benefit of others): using `devtools::document()` on the package leads R check to throw an error because RcppArmadillo (and Rcpp::sourceEval) are added to the NAMESPACE but RcppArmadillo is not in the DESCRIPTION. Thus you have been using `devtools::documents(roclets = 'rd')`, i.e. not using roxygen2 to generate the NAMESPACE. If I understand correctly, those two imports are not needed, and thus the preferable solution is to remove the directives in `mfsurv.R` that tell roxygen2 to import them.
- Re testing: this was more about the **Automated tests** item in our review checklists. The idea is to make it easier to catch issues like that early. Like I said, I don't think you need full blown unit testing with `testthat`, but since it shouldn't be that much work to add Travis or some other CI, could you please go ahead and do it?
