### 2020-03-24 59f6621
errorRcppArmadillo
