+++ NAMESPACE
@@ -12,7 +12,6 @@ import(graphics)
 importFrom(FastGP,rcpp_log_dmvnorm)
 importFrom(FastGP,rcpp_rmvnorm)
 importFrom(MCMCpack,riwish)
-importFrom(Rcpp,evalCpp)
 importFrom(coda,mcmc)
 importFrom(mvtnorm,dmvnorm)
 importFrom(mvtnorm,rmvnorm)
