### 2020-03-24 Nicolas-Schmidt/BayesMFSurv#1 @Nicolas-Schmidt
Hi @andybega!

Oh, you're right!
Until now we have not realized that it was a comment from roxygen2 (# ') instead of a simple comment (#). We have already fixed it.

Thank you!
