### 2020-03-24 1e51d81
rmerrorRcppArmadillo
