### 2020-03-24 #2164 @Sergio-Bejar
Hi @andybega,

You are absolutely right!   We somehow missed those points, and we apologize for that.

We have (i) removed the directives in `mfsurv.R` that tell roxygen2 to import them, and (ii) added Travis as suggested by you.
