### 2020-03-28 #2164 @alletsee
Hi all.

Thanks for taking the comments into account. I now found the time to re-read the article and run basic checks on functionality. From my perspective, all aspects mentioned above have been addressed thoroughly.

@Nicolas-Schmidt @Sergio-Bejar : Thank you for the effort of making this publicly available.

@marcosvital : my review is also complete.
