### 2020-03-30 #2164 @marcosvital
Good, all DOIs are ok!

@Nicolas-Schmidt and @Sergio-Bejar , I'll generate a (possibly) final proof. Please check if everything is ok with it. I'll read it myself too, and will let you know if I find something that should be fixed.
