### 2020-03-30 #2164 @Sergio-Bejar
@marcosvital , I made a couple of minor changes in the paper (nothing major, please see proof generated above).  Please let me know if there's anything you'd like to change/add.

Thanks!
