### 2020-04-09 #2164 @marcosvital
Hello again, @Sergio-Bejar and @Nicolas-Schmidt, and sorry about the delay.

Everything looks fine for me, so we can move on to acceptance!

As one of our last steps, if you already didn't do this, you will need to archive the last release of the package (on Zenodo, figshare, or other of your choice). After that, post the version number and archive DOI here, and we can continue.
