### 2020-04-09 #2164 @Nicolas-Schmidt
Hi @marcosvital , thanks for everything!
The package version is v0.2.0.
DOI is https://zenodo.org/badge/latestdoi/226979997
