+++ README.md
@@ -8,6 +8,8 @@ Yadav*
 
 <!-- badges: start -->
 
+[![R build
+status](https://github.com/Nicolas-Schmidt/BayesMFSurv/workflows/R-CMD-check/badge.svg)](https://github.com/Nicolas-Schmidt/BayesMFSurv/actions)
 [![CRAN\_Status\_Badge](https://www.r-pkg.org/badges/version/BayesMFSurv)](https://cran.r-project.org/package=BayesMFSurv)
 [![Project Status: Active â€“ The project has reached a stable, usable
 state and is being actively
