### 2020-04-24 8b1aa1c
minor updates to paper
