### 2020-04-24 #2164 @danielskatz
@Nicolas-Schmidt - please merge https://github.com/Nicolas-Schmidt/BayesMFSurv/pull/6
